rename `client/src/components/Workflow/test-utils.ts` to `testUtils`
